refactor: rename reporters according to Airbnb style guide
