+++ src/reporters/ApiaryReporter.js
