+++ src/reporters/BaseReporter.js
