+++ src/reporters/CliReporter.js
