+++ src/reporters/DotReporter.js
