+++ src/reporters/NyanReporter.js
