+++ src/reporters/XUnitReporter.js
