+++ test/unit/reporters/NyanReporter-test.js
@@ -6,7 +6,7 @@ const { EventEmitter } = require('events');
 
 const loggerStub = require('../../../src/logger');
 
-const NyanCatReporter = proxyquire('../../../src/reporters/nyan-reporter', {
+const NyanCatReporter = proxyquire('../../../src/reporters/NyanReporter', {
   '../logger': loggerStub
 });
